refactor : Line객체에서 Sections 객체 생성 및 비즈니스 로직 이동
